Renamed some files / classes to fix Sonarqube code smells
