+++ frontend/src/devicelist/StatusIndicator.js
